requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
numpy==1.20.0
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
z3-solver==4.8.10.0
